Added VMCALL/hypervisor module within plugin-loader
This will give us an easy interface directly with the hypervisor.
Currently the only action implemented is the check_presence one.

This HAS NOT BEEN TESTED, I've probably barfed the inline assembly
for the VMCALL or some other logic related to VEH. TEST BEFORE
GOING ANY FURTHER!

+++ interface/vmcall_if.h
@@ -19,7 +19,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-#define VMCALL_SECRET_KEY 0x0CA5CADE
+#define VMCALL_SECRET_KEY ((size_t)0x0CA5CADE)
 
 enum vmcall_action {
     ACTION_CHECK_PRESENCE,
